update place selector to have default method of getting child place options
