+++ static/js/shared/place_selector.tsx
@@ -22,23 +22,62 @@ import _ from "lodash";
 import React, { useEffect, useState } from "react";
 import { Card, Container, CustomInput } from "reactstrap";
 
+import { isChildPlaceOf } from "../tools/shared_util";
 import {
   ENCLOSED_PLACE_TYPE_NAMES,
   getNamedTypedPlace,
   getParentPlacesPromise,
 } from "../utils/place_utils";
+import { EARTH_NAMED_TYPED_PLACE, USA_PLACE_DCID } from "./constants";
 import { SearchBar } from "./place_search_bar";
 import { NamedTypedPlace } from "./types";
 
 const EMPTY_NAMED_TYPED_PLACE = { dcid: "", name: "", types: null };
 const SELECTOR_PREFIX = "place-selector";
+const USA_CITY_CHILD_TYPES = ["CensusZipCodeTabulationArea", "City"];
+const USA_COUNTY_CHILD_TYPES = ["Town", "Village", ...USA_CITY_CHILD_TYPES];
+const USA_STATE_CHILD_TYPES = ["County", ...USA_COUNTY_CHILD_TYPES];
+const USA_COUNTRY_CHILD_TYPES = ["State", ...USA_STATE_CHILD_TYPES];
 
+const USA_CHILD_PLACE_TYPES = {
+  Country: USA_COUNTRY_CHILD_TYPES,
+  State: USA_STATE_CHILD_TYPES,
+  County: USA_COUNTY_CHILD_TYPES,
+  City: USA_CITY_CHILD_TYPES,
+};
+
+const AA4_CHILD_PLACE_TYPES = ["AdministrativeArea5"];
+const AA3_CHILD_PLACE_TYPES = ["AdministrativeArea4", ...AA4_CHILD_PLACE_TYPES];
+const AA2_CHILD_PLACE_TYPES = ["AdministrativeArea3", ...AA3_CHILD_PLACE_TYPES];
+const AA1_CHILD_PLACE_TYPES = ["AdministrativeArea2", ...AA2_CHILD_PLACE_TYPES];
+const NUTS2_CHILD_PLACE_TYPES = ["EurostatNUTS3"];
+const NUTS1_CHILD_PLACE_TYPES = ["EurostatNUTS2", ...NUTS2_CHILD_PLACE_TYPES];
+const NON_USA_COUNTRY_PLACE_TYPES = [
+  "AdministrativeArea1",
+  ...AA1_CHILD_PLACE_TYPES,
+  "EurostatNUTS1",
+  ...NUTS1_CHILD_PLACE_TYPES,
+];
+const CONTINENT_PLACE_TYPES = ["Country", ...NON_USA_COUNTRY_PLACE_TYPES];
+const CHILD_PLACE_TYPES = {
+  Planet: ["Continent", ...CONTINENT_PLACE_TYPES, ...USA_COUNTRY_CHILD_TYPES],
+  Continent: CONTINENT_PLACE_TYPES,
+  Country: NON_USA_COUNTRY_PLACE_TYPES,
+  State: AA1_CHILD_PLACE_TYPES,
+  EurostatNUTS1: NUTS1_CHILD_PLACE_TYPES,
+  EurostatNUTS2: NUTS2_CHILD_PLACE_TYPES,
+  AdministrativeArea1: AA1_CHILD_PLACE_TYPES,
+  AdministrativeArea2: AA2_CHILD_PLACE_TYPES,
+  AdministrativeArea3: AA3_CHILD_PLACE_TYPES,
+  AdministrativeArea4: AA4_CHILD_PLACE_TYPES,
+};
 interface PlaceSelectorProps {
   selectedPlace: NamedTypedPlace;
   enclosedPlaceType: string;
   onPlaceSelected: (place: NamedTypedPlace) => void;
   onEnclosedPlaceTypeSelected: (type: string) => void;
-  getEnclosedPlaceTypes: (
+  // special handling of how to get enclosed place types
+  getEnclosedPlaceTypes?: (
     place: NamedTypedPlace,
     parentPlaces: NamedTypedPlace[]
   ) => string[];
@@ -58,44 +97,46 @@ export function PlaceSelector(props: PlaceSelectorProps): JSX.Element {
   return (
     <Card className={`${SELECTOR_PREFIX}-card`}>
       <Container className={`${SELECTOR_PREFIX}-container`} fluid={true}>
-        <div
-          className={`${SELECTOR_PREFIX}-section`}
-          id={`${SELECTOR_PREFIX}-search-section`}
-        >
-          <div className={`${SELECTOR_PREFIX}-label`}>Plot places in</div>
-          <SearchBar
-            places={
-              props.selectedPlace.dcid
-                ? { [props.selectedPlace.dcid]: props.selectedPlace.name }
-                : {}
-            }
-            addPlace={(e) => selectPlace(e, props.onPlaceSelected)}
-            removePlace={() =>
-              unselectPlace(props.onPlaceSelected, setChildPlaceTypes)
-            }
-            numPlacesLimit={1}
-            customPlaceHolder={props.customSearchPlaceholder}
-          />
-        </div>
-        <div className={`${SELECTOR_PREFIX}-section`}>
-          <div className={`${SELECTOR_PREFIX}-label`}>of type</div>
-          <div>
-            <CustomInput
-              id={`${SELECTOR_PREFIX}-place-type`}
-              type="select"
-              value={props.enclosedPlaceType}
-              onChange={(e) =>
-                selectEnclosedPlaceType(e, props.onEnclosedPlaceTypeSelected)
+        <div className={`${SELECTOR_PREFIX}-main-selectors`}>
+          <div
+            className={`${SELECTOR_PREFIX}-section`}
+            id={`${SELECTOR_PREFIX}-search-section`}
+          >
+            <div className={`${SELECTOR_PREFIX}-label`}>Plot places in</div>
+            <SearchBar
+              places={
+                props.selectedPlace.dcid
+                  ? { [props.selectedPlace.dcid]: props.selectedPlace.name }
+                  : {}
               }
-              className="pac-target-input"
-            >
-              <option value="">Select a place type</option>
-              {childPlaceTypes.map((type) => (
-                <option value={type} key={type}>
-                  {ENCLOSED_PLACE_TYPE_NAMES[type] || type}
-                </option>
-              ))}
-            </CustomInput>
+              addPlace={(e) => selectPlace(e, props.onPlaceSelected)}
+              removePlace={() =>
+                unselectPlace(props.onPlaceSelected, setChildPlaceTypes)
+              }
+              numPlacesLimit={1}
+              customPlaceHolder={props.customSearchPlaceholder}
+            />
+          </div>
+          <div className={`${SELECTOR_PREFIX}-section`}>
+            <div className={`${SELECTOR_PREFIX}-label`}>of type</div>
+            <div>
+              <CustomInput
+                id={`${SELECTOR_PREFIX}-place-type`}
+                type="select"
+                value={props.enclosedPlaceType}
+                onChange={(e) =>
+                  selectEnclosedPlaceType(e, props.onEnclosedPlaceTypeSelected)
+                }
+                className="pac-target-input"
+              >
+                <option value="">Select a place type</option>
+                {childPlaceTypes.map((type) => (
+                  <option value={type} key={type}>
+                    {ENCLOSED_PLACE_TYPE_NAMES[type] || type}
+                  </option>
+                ))}
+              </CustomInput>
+            </div>
           </div>
         </div>
         {props.children}
@@ -106,7 +147,9 @@ export function PlaceSelector(props: PlaceSelectorProps): JSX.Element {
   function loadChildPlaceTypes(): void {
     getParentPlacesPromise(props.selectedPlace.dcid)
       .then((parentPlaces) => {
-        const newChildPlaceTypes = props.getEnclosedPlaceTypes(
+        const getEnclosedPlaceTypesFn =
+          props.getEnclosedPlaceTypes || getEnclosedPlaceTypes;
+        const newChildPlaceTypes = getEnclosedPlaceTypesFn(
           props.selectedPlace,
           parentPlaces
         );
@@ -149,4 +192,29 @@ function selectEnclosedPlaceType(
 ): void {
   const placeType = event.target.value;
   onEnclosedPlaceTypeSelected(placeType);
+}
+
+function getEnclosedPlaceTypes(
+  place: NamedTypedPlace,
+  parentPlaces: NamedTypedPlace[]
+): string[] {
+  if (place.dcid === EARTH_NAMED_TYPED_PLACE.dcid) {
+    return CHILD_PLACE_TYPES[EARTH_NAMED_TYPED_PLACE.types[0]];
+  }
+  if (_.isEmpty(place.types)) {
+    return [];
+  }
+  const isUSPlace = isChildPlaceOf(place.dcid, USA_PLACE_DCID, parentPlaces);
+  for (const type of place.types) {
+    if (isUSPlace) {
+      if (type in USA_CHILD_PLACE_TYPES) {
+        return USA_CHILD_PLACE_TYPES[type];
+      }
+    } else {
+      if (type in CHILD_PLACE_TYPES) {
+        return CHILD_PLACE_TYPES[type];
+      }
+    }
+  }
+  return [];
 }
